# openapi.yaml
openapi: 3.0.0
info:
  title: PetStore MCP Demo
  version: 1.0.0
paths:
  /pets:
    get:
      summary: List all pets
      parameters:
        - in: query
          name: tag
          schema:
            type: string
          description: Filter pets by tag (case-insensitive)
      responses:
        '200':
          description: A list of pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
    post:
      summary: Create a pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPet'
      responses:
        '201':
          description: Pet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '400':
          description: Invalid input
  /pets/{id}:
    get:
      summary: Get a single pet
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pet found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '404':
          description: Pet not found
    delete:
      summary: Delete a pet
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Pet deleted
        '404':
          description: Pet not found

components:
  schemas:
    Pet:
      type: object
      required: [id, name, tag]
      properties:
        id:
          type: integer
        name:
          type: string
        tag:
          type: string
    NewPet:
      type: object
      required: [name, tag]
      properties:
        name:
          type: string
        tag:
          type: string